Write to Route Table/Daily Drivers
Writing to Route table is complete and seems to be bug free.
Drivers available each day are now being found and they are being assigned to the routes that get generated.
The amount of clusters/routes that get created are based off of how many drivers are available that day.